*Particle system initial implement.
*Add color GLubyte function.

*Arrange GameObject system.

+++ Engine/Engine/Vector2.cpp
@@ -1,12 +1,13 @@
+
 #include "Vector2.h"
 
 Vector2::Vector2()
 {
-	x = 0;
-	y = 0;
+	x = 0.0f;
+	y = 0.0f;
 }
 
-Vector2::Vector2(int x, int y)
+Vector2::Vector2(float x, float y)
 {
 	this->x = x;
 	this->y = y;
